\name{changeLabel}
\alias{changeLabel}
\title{changeLabel}
\description{Describe an annotated region label for supervised change-point detection.}
\usage{changeLabel(annotation, 
    min.changes, max.changes, 
    color)}
\arguments{
  \item{annotation}{annotation }
  \item{min.changes}{min.changes }
  \item{max.changes}{max.changes }
  \item{color}{color }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





