\name{check_features_targets}
\alias{check_features_targets}
\title{check features targets}
\description{stop with an informative error if there is a problem with the
feature or target matrix.}
\usage{check_features_targets(feature.mat, 
    target.mat)}
\arguments{
  \item{feature.mat}{n x p numeric input feature matrix.}
  \item{target.mat}{n x 2 matrix of target interval limits.}
}

\value{number of observations/rows.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





