\name{check_target_pred}
\alias{check_target_pred}
\title{check target pred}
\description{stop with an informative error if there are problems with the
target matrix or predicted values.}
\usage{check_target_pred(target.mat, pred)}
\arguments{
  \item{target.mat}{
}
  \item{pred}{
}
}

\value{number of observations.}

\author{Toby Dylan Hocking}





