\name{demo8}
\alias{demo8}
\docType{data}
\title{
  PeakSegFPOP demo data set
}
\description{
  PeakSegFPOP demo data set with 8 observations
}
\usage{data("demo8")}
\format{
  A list of two objects: feature.mat is an 8 x 36 input feature matrix,
  and target.mat is a 8 x 2 output limit matrix.
}
\keyword{datasets}
