\name{geom_tallrect}
\alias{geom_tallrect}
\title{geom tallrect}
\description{ggplot2 geom with xmin and xmax aesthetics that covers the entire
y range, useful for clickSelects background elements.}
\usage{geom_tallrect(mapping = NULL, 
    data = NULL, stat = "identity", 
    position = "identity", 
    ..., na.rm = FALSE, 
    show.legend = NA, 
    inherit.aes = TRUE)}
\arguments{
  \item{mapping}{mapping }
  \item{data}{data }
  \item{stat}{stat }
  \item{position}{position }
  \item{\dots}{\dots }
  \item{na.rm}{na.rm }
  \item{show.legend}{show.legend }
  \item{inherit.aes}{inherit.aes }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





