\name{IntervalRegressionUnregularized}
\alias{IntervalRegressionUnregularized}
\title{IntervalRegressionUnregularized}
\description{Use IntervalRegressionRegularized with initial.regularization=0
and factor.regularization=NULL, meaning fit one un-regularized
interval regression model.}
\usage{IntervalRegressionUnregularized(...)}
\arguments{
  \item{\dots}{passed to IntervalRegressionRegularized.}
}

\value{List representing fit model, see
help(IntervalRegressionRegularized) for details.}

\author{Toby Dylan Hocking}





