\name{notConverging}
\alias{notConverging}
\docType{data}
\title{
  Interval regression problem that was not converging
}
\description{
  A small data set which was diverging using a previous implementation
  of IntervalRegressionCV.
}
\usage{data("notConverging")}
\format{
  A list with names: X.mat are numeric inputs, y.mat are numeric
  outputs, fold.vec is an integer vector of fold ID numbers.
}
\source{
  github.com/tdhock/neuroblastoma-data, data/H3K4me3_TDH_other/cv/equal_labels/testFolds/3/sampleSelectionGP_erf/5/order.csv
}
\keyword{datasets}
