\name{oneSkip}
\alias{oneSkip}
\docType{data}
\title{
oneSkip
}
\description{
  A loss and model complexity function which never selects one of the
  models, using a linear penalty.
}
\usage{data("oneSkip")}
\format{
  A list of two data.frames (input and output).
}
\source{
  example(exactModelSelection) in PeakSegDP package.
}
\keyword{datasets}
