\name{plot.IntervalRegression}
\alias{plot.IntervalRegression}
\title{plot IntervalRegression}
\description{Plot an IntervalRegression model.}
\usage{\method{plot}{IntervalRegression}(x, 
    ...)}
\arguments{
  \item{x}{x }
  \item{\dots}{\dots }
}

\value{a ggplot.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





