\name{predict.IntervalRegression}
\alias{predict.IntervalRegression}
\title{predict IntervalRegression}
\description{Compute model predictions.}
\usage{\method{predict}{IntervalRegression}(object, X, ...)}
\arguments{
  \item{object}{
}
  \item{X}{
}
  \item{\dots}{
}
}

\value{numeric matrix of predicted log(penalty) values.}

\author{Toby Dylan Hocking}





