\name{print.IntervalRegression}
\alias{print.IntervalRegression}
\title{print IntervalRegression}
\description{print learned model parameters.}
\usage{\method{print}{IntervalRegression}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}

\value{numeric matrix of learned parameters (on the original feature
scale), can be used for prediction via cbind(1,features) \%*\%
parameters.}

\author{Toby Dylan Hocking}





