\name{changeLabel}
\alias{changeLabel}
\title{changeLabel}
\description{Describe an annotated region label for supervised change-point detection.}
\usage{changeLabel(annotation, min.changes, max.changes, color)}
\arguments{
  \item{annotation}{
}
  \item{min.changes}{
}
  \item{max.changes}{
}
  \item{color}{
}
}



\author{Toby Dylan Hocking}





