\name{IntervalRegressionUnregularized}
\alias{IntervalRegressionUnregularized}
\title{IntervalRegressionUnregularized}
\description{Use \code{\link{IntervalRegressionRegularized}} with initial.regularization=0
and factor.regularization=NULL, meaning fit one un-regularized
interval regression model.}
\usage{IntervalRegressionUnregularized(...)}
\arguments{
  \item{\dots}{passed to \code{\link{IntervalRegressionRegularized}}.}
}

\value{List representing fit model, see
help(\code{\link{IntervalRegressionRegularized}}) for details.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





