\name{predict.IntervalRegression}
\alias{predict.IntervalRegression}
\title{predict IntervalRegression}
\description{Compute model predictions.}
\usage{\method{predict}{IntervalRegression}(object, 
    X, ...)}
\arguments{
  \item{object}{object }
  \item{X}{X }
  \item{\dots}{\dots }
}

\value{numeric matrix of predicted log(penalty) values.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





